+++ src/lib.rs
@@ -89,7 +89,22 @@ pub mod adc;
 pub mod bb;
 #[cfg(all(
     feature = "can",
-    any(feature = "stm32f405", feature = "stm32f407", feature = "stm32f446")
+    any(
+        feature = "stm32f405",
+        feature = "stm32f407",
+        feature = "stm32f412",
+        feature = "stm32f413",
+        feature = "stm32f415",
+        feature = "stm32f417",
+        feature = "stm32f423",
+        feature = "stm32f427",
+        feature = "stm32f429",
+        feature = "stm32f437",
+        feature = "stm32f439",
+        feature = "stm32f446",
+        feature = "stm32f469",
+        feature = "stm32f479",
+    )
 ))]
 pub mod can;
 #[cfg(feature = "device-selected")]